requirements.txt
fastapi
uvicorn
requests
python-dotenv
gspread
oauth2client
